requirements.txt
fastapi
dotenv
uvicorn
python-telegram-bot
httpx